requirements/_base.txt
#distribute==0.6.24
Django==1.6.5
South==0.8.4
argparse==1.2.1
cssutils==0.9.10b1
django-filter==0.7
django-floppyforms==1.2.0
django-grappelli==2.5.3
django-leaflet==0.13.2
django-model-utils==2.0.3
django-post_office==0.8.4
django-social-auth==0.7.28
django-widget-tweaks==1.3
djangorestframework==2.3.14
dnspython==1.10.0
jsonfield==0.9.20
lxml==3.0.1
mimeparse==0.1.3
psycopg2==2.4.5
pynliner==0.4.0
python-dateutil==1.5
wsgiref==0.1.2
